Made button go 'Hello World' using javascript
Ran into small problem trying to get a button to make an alert appear on
the page that says 'Hello World'. The problem was the script was trying
to access the button element but the script itself was being loaded
before the document content. So moved the <script> tag to the bottom of
the document just before the </body> tag. This made the script aware of
the button in the DOM